+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.32.5 (July 09, 2019)
+
+- feat: support config forge chain/node from forge-cli #86
+- fix: `forge --version` is broken
+- feat: support creating wallets in forge-cli without a running node #4
+- chore: update readme
+- chore: update update readme script
+- chore: update `forge stop` description
+
 ## 0.32.4 (July 08, 2019)
 
 - feat: display account type info when do `forge account xxx` #74